moved ObjectComparisonTests into test project
